# PR Summary
**Title:** Refactor: Modularize `extract_source_mappings.py` for improved maintainability

### Context

The `tritonparse/tritonparse/extract_source_mappings.py` script was a large, monolithic file containing logic for IR parsing, source mapping, event diffing, and file processing. This made it difficult to read, maintain, and extend.

This pull request refactors the script by breaking it down into smaller, single-responsibility modules. This improves the overall code structure, enhances readability, and makes future development easier.

### Description of Changes

The core logic of `extract_source_mappings.py` has been split into several new modules within the `tritonparse/tritonparse/` directory:

*   **`extract_source_mappings.py` (Modified):** Now serves as a clean command-line entry point, containing only argument parsing and the main execution call.
*   **`trace_processor.py` (New, formerly `processor.py`):** The main orchestrator that handles the processing of trace files. It contains the core logic previously found in `parse_single_file` and `parse_single_trace_content`.
*   **`ir_parser.py` (New):** Contains all functions related to parsing Intermediate Representations (TTIR, TTGIR, PTX, AMDGCN), including `loc` directive extraction.
*   **`mapper.py` (New):** Responsible for creating the bidirectional source mappings between Python source and various IRs.
*   **`event_diff.py` (New, formerly `launch_diff.py`):** Provides functionality to compare a list of events and generate a summary of their differences.
*   **`sourcemap_utils.py` (New):** A collection of general-purpose utility functions (e.g., dictionary flattening) used by the new modules.

Additionally, the `sourcemap_constants.py` file was **removed**, and its contents (regex patterns and constant lists) were moved into the modules where they are directly used (`ir_parser.py` and `event_diff.py`) to reduce the number of files and improve cohesion.
